Use native int for OCL workgroup sizes in Gaussian blur
Simplify the OpenCL workgroup size computation by using int instead
of size_t, avoiding unnecessary type widening on 32-bit platforms.
This aligns with how the values are actually consumed downstream.

+++ modules/imgproc/src/smooth.dispatch.cpp
@@ -338,8 +338,8 @@ static bool ocl_GaussianBlur_8UC1(InputArray _src, OutputArray _dst, Size ksize,
         ddepth = sdepth;
 
     Size size = _src.size();
-    size_t globalsize[2] = { 0, 0 };
-    size_t localsize[2] = { 0, 0 };
+    int globalsize[2] = { 0, 0 };
+    int localsize[2] = { 0, 0 };
 
     if (ksize.width == 3)
     {
